cloud-batch-python/requirements.txt
absl-py==1.3.0
ansible==7.0.0
ansible-core==2.13.4
apache-libcloud==3.4.1
apparmor==3.0.7
argcomplete==2.0.0
asttokens==2.0.5
attrs==22.1.0
Babel==2.10.3
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.11.1
blinker==1.5
Brlapi==0.8.4
cached-property==1.5.2
cachetools==5.2.0
certifi==2022.9.24
chardet==4.0.0
charset-normalizer==3.0.1
chrome-gnome-shell==0.0.0
credentialkit-client==1
crudini==0.9.3
cryptography==3.4.8
cssselect==1.2.0
cupshelpers==1.0
dasbus==1.7
dbus-python==1.3.2
debugpy==1.6.2
decorator==5.1.1
distro==1.8.0
distro-info==1.2
dnspython==2.2.1
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
duplicity==0.8.22
entrypoints==0.4
executing==0.9.1
fasteners==0.17.3
future==0.18.2
gbulb==0.6.1
glinux-boot-image-kthresher==1.4.1
glinux-homedir-helper==1
glinux-identity==1
glinux-shutdown-installer==1.0
glinux-welcome==32
goobuntu-config-tools==0.1
goobuntu-gce-login-policy==0.1
goobuntu-sso-watcher==0.1
google-api-core==2.10.2
google-auth==2.12.0
google-cloud==0.34.0
google-cloud-batch==0.4.0
googleapis-common-protos==1.56.4
gpg==1.18.0
grpc-google-iam-v1==0.12.4
grpcio==1.50.0
grpcio-status==1.49.1
gyp==0.1
hg-evolve==10.6.0.dev0
html5lib==1.1
httplib2==0.20.4
idna==3.3
importlib-metadata==4.12.0
iniparse==0.5
invoke==1.7.0
IPy==1.1
ipykernel==6.15.1
ipython==8.4.0
ipywidgets==8.0.1
jedi==0.18.1
Jinja2==3.0.3
jmespath==1.0.1
jsonpointer==2.3
jsonschema==4.7.2
jupyter-client==7.3.4
jupyter-core==4.11.1
jupyterlab-widgets==3.0.1
kernel-pruner==60
LibAppArmor==3.0.7
libevdev==0.5
lockfile==0.12.2
louis==3.23.0
lxml==4.9.1
lz4==4.0.2+dfsg
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mercurial==6.2.2+hg61.d513ae93dff3
monotonic==1.6
more-itertools==8.10.0
nest-asyncio==1.5.5
netaddr==0.8.0
ntlm-auth==1.4.0
numpy==1.21.5
oauthlib==3.2.1
obno==46
olefile==0.46
onboard==1.4.1
packaging==21.3
PAM==0.4.2
paramiko==2.10.4
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
prompt-toolkit==3.0.30
proto-plus==1.22.1
protobuf==4.21.7
psutil==5.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.20.1
pycryptodomex==3.11.0
pycups==2.0.1
pycurl==7.45.1
Pygments==2.13.0
PyGObject==3.42.2
pyinotify==0.9.6
PyJWT==2.4.0
PyKCS11==1.5.11
pykerberos==1.1.14
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyparsing==3.0.7
pyrsistent==0.18.1
pysmbc==1.0.23
python-apt==2.5.0
python-augeas==0.5.0
python-dateutil==2.8.2
python-debian==0.1.49
python-dotenv==0.21.0
python-networkmanager==2.2
python-pam==2.0.2
python-xlib==0.31
pytz==2022.6
pyudev==0.24.0
pywinrm==0.3.0
pyxattr==0.7.2
pyxdg==0.27
PyYAML==6.0
pyzmq==23.2.0
reboot-enforcer==0.1
reconfigure==0.1.81
rekey==1
requests==2.28.1
requests-kerberos==0.12.0
requests-ntlm==1.1.0
requests-toolbelt==0.9.1
resolvelib==0.8.1
rfc3987==1.3.8
rsa==4.9
scour==0.38.2
selinux==3.4
setproctitle==1.3.1
simplejson==3.17.6
six==1.16.0
soupsieve==2.3.2
stack-data==0.3.0
texttable==1.6.4
tinycss==0.4
tinycss2==1.2.1
tornado==6.2
traitlets==5.3.0
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.26.12
wcwidth==0.2.5
webcolors==1.11.1
webencodings==0.5.1
websocket-client==1.2.3
widgetsnbextension==4.0.1
xdg==5
xlrd==1.2.0
xmltodict==0.13.0
youtube-dl==2021.12.17
zipp==1.0.0
